restructure a bit
