+++ src/app.scss
@@ -1,6 +1,6 @@
 @import '~@nio/ui-kit/dist/nio';
-@import './css/vars';
-@import './css/base';
+@import './assets/css/vars';
+@import './assets/css/base';
 
 /*    place your app-specific css below (or make component/view   */
 /*    specific scss files and throw them in the css folder)       */
